+++ esp_data/__init__.py
@@ -1,0 +1,17 @@
+from .dataset import (
+    Dataset,
+    DatasetInfo,
+    dataset_from_config,
+    list_registered_datasets,
+    print_registered_datasets,
+    register_dataset,
+)
+
+__all__ = [
+    "dataset_from_config",
+    "Dataset",
+    "DatasetInfo",
+    "list_registered_datasets",
+    "print_registered_datasets",
+    "register_dataset",
+]
