+++ esp_data/datasets/__init__.py
@@ -1,21 +1,7 @@
 from .animalspeak import AnimalSpeak
 from .barkley_canyon import BarkleyCanyon
-from .base import (
-    Dataset,
-    DatasetInfo,
-    dataset_from_config,
-    list_registered_datasets,
-    print_registered_datasets,
-    register_dataset,
-)
 
 __all__ = [
     "AnimalSpeak",
     "BarkleyCanyon",
-    "dataset_from_config",
-    "Dataset",
-    "DatasetInfo",
-    "list_registered_datasets",
-    "print_registered_datasets",
-    "register_dataset",
 ]
