+++ pyproject.toml
@@ -12,6 +12,7 @@ dependencies = [
     "librosa>=0.11.0",
     "pandas",
     "pydantic>=2",
+    "resampy",
     "s3fs>=2024.9.0",
     "semver",
     "soundfile>=0.13.1",
