+++ requirements.txt
@@ -105,6 +105,7 @@ pyyaml-env-tag==0.1
 requests==2.32.3
 requests-oauthlib==2.0.0
 requirements-parser==0.11.0
+resampy==0.4.3
 rsa==4.9.1
 ruff==0.11.8
 s3fs==2024.12.0
