+++ tests/test_barkley_canyon.py
@@ -2,8 +2,10 @@
 
 import pytest
 
-from esp_data_temp.config import DatasetConfig
-from esp_data_temp.datasets import BarkleyCanyon, Dataset
+from esp_data.config import DatasetConfig
+from esp_data.datasets import BarkleyCanyon
+from esp_data import Dataset
+
 
 
 @pytest.fixture
